Refactored condition UI into Conditions class
